runoff.md and tideman.md complete
I cleaned up both runoff.md and tideman.md  with the aim of making them more readable. I have also included a table of contents for each of the sections of each md with their perspective line numbers in the file. My hope is that by doing so, the user can go to a specific section if they have something specific in mind.